Jay Benesch edit increment four   Did not address Jay's stored energy comment:
Why do you keep writing of "stored energy" in a vacuum system?
The energy source is the atmosphere, not anything stored in the vacuum.

I suppose one could say that energy was stored in the vacuum via the
action of the pumps which removed the air.

I'd prefer to say "Air pressure is ~1 kg/cm^2.  If the integrity of a
large vacuum window is violated the air inrush may cause damage to
hearing and will turn the window into shrapnel which could injure
nearby personnel."